refactor: move the test to separate test file
